migrate new database from user2 to user
